+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.528.3-4] - 2026-01-21
+### Added
+- [#241] Integration for Kubernetes-native Jenkins agents.
+
 ## [v2.528.3-3] - 2026-01-15
 ### Changed
 - [#244] Localization and packet handling improvements